move utils to own file
