+++ packages/accordion/docs/overview.mdx
@@ -2,9 +2,8 @@
 title: Accordion
 description: An accordion shows or hides sections of content. It lets users see only the content they need.
 group: Layout
+storybookPath: /story/layout-accordion--on-light
 ---
-
-<StoryBookPreview src="https://steelthreads.github.io/agds-next/storybook/iframe.html?args=&id=layout-accordion--on-light&viewMode=story" />
 
 Users can open or close an accordion to decide the content they see.
