+++ packages/button/docs/overview.mdx
@@ -2,9 +2,8 @@
 title: Button
 description: A button makes an action obvious to the user, as well as making an action faster and easier to perform.
 group: Forms
+storybookPath: /story/forms-button--on-light
 ---
-
-<StoryBookPreview src="https://steelthreads.github.io/agds-next/storybook/iframe.html?args=&id=forms-button--on-light&viewMode=story" />
 
 A button indicates to a user what action will happen next.
