--- templates/index.mdx
@@ -1,1 +1,0 @@
-
